requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.3.3
idna==2.9
multidict==4.7.5
praw==6.5.1
prawcore==1.0.1
pycodestyle==2.5.0
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
update-checker==0.16
urllib3==1.25.9
websocket-client==0.57.0
websockets==8.1
yarl==1.4.2
